feat: dynamic import
- 클라이언드 api인 navigator에 접근하기 때문에 Dynamic import로 변경
